requirements.txt
sqlalchemy>=1.3.10
databases
asyncpg
psycopg2-binary
